fix: AudioEffectRecord::get_recording() crashes from time to time
The size of recording_data may change DURING the call to get_recording(),
because another thread may write to it.
Big kudos to @Adrenesis for sweating this one out.

This commit deploys a Mutex to force sequential access to recording_data.

Not sure about the usage of ifndef in the header.
Pretty bad fix, since it makes get_recording() blocking.