Fix AudioEffectRecord::get_recording() crashes from time to time
Thx to @Goutte for C++ help =)

A bit of cleanup.

fix: Mutex is created in the wrong place. by @Goutte

(merge blooper)

Make AudioEffectRecord encode on its thread.

fix: move mutex creation to class constructor by @Goutte

Kudos to @Adrenesis again for finding the crash on windows.

fix: AudioEffectRecord::get_recording() crashes from time to time by @Goutte

The size of recording_data may change DURING the call to get_recording(),
because another thread may write to it.
Big kudos to @Adrenesis for sweating this one out.

This commit deploys a Mutex to force sequential access to recording_data. by @Goutte

Not sure about the impact on sound quality of this fix.
Not sure about the usage of ifndef in the header.

fix: AudioEffectRecord::get_recording() crashes from time to time by @Goutte

The size of recording_data may change DURING the call to get_recording(),
because another thread may write to it.
Big kudos to @Adrenesis for sweating this one out.